Revise Controller

+++ src/main/java/com/homework/session/service/CommentService.java
@@ -9,10 +9,12 @@ import com.homework.session.entity.BoardList;
 import com.homework.session.entity.Comment;
 import com.homework.session.entity.User;
 import com.homework.session.error.exception.UnAuthorizedException;
+import com.homework.session.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,12 +27,19 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final BoardRepository boardRepository;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Transactional
-    public Long createComment(Long id, String nickname, CommentRequestDto commentDto) {
-        User user = userRepository.findByNickname(nickname);
+    public Long createComment(Long id, CommentRequestDto commentDto, HttpServletRequest request) {
+
+        String token = jwtTokenProvider.resolveAccessToken(request);
+        String email = jwtTokenProvider.getUserEmail(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
         BoardList boardList = boardRepository.findById(id).orElseThrow(() ->
-                new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION));
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
 
         commentDto.setUser(user);
         commentDto.setBoardList(boardList);
@@ -50,17 +59,39 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(Long id, CommentRequestDto commentRequestDto) {
+    public void updateComment(Long id, CommentRequestDto commentRequestDto, HttpServletRequest request) {
+
+        String token = jwtTokenProvider.resolveAccessToken(request);
+        String email = jwtTokenProvider.getUserEmail(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
         Comment comment = commentRepository.findById(id).orElseThrow(() ->
-                new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION));
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
+        if (!comment.getUser().getNickname().equals(user.getNickname())) {
+            throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION);
+        }
 
         comment.update(commentRequestDto.getComment());
     }
 
     @Transactional
-    public void deleteComment(Long id) {
+    public void deleteComment(Long id, HttpServletRequest request) {
+
+        String token = jwtTokenProvider.resolveAccessToken(request);
+        String email = jwtTokenProvider.getUserEmail(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
         Comment comment = commentRepository.findById(id).orElseThrow(() ->
-                new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION));
+        { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
+        if (!comment.getUser().getNickname().equals(user.getNickname())) {
+            throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION);
+        }
 
         commentRepository.delete(comment);
     }
